MotionControl Dirve straight first attempt it compiles
Created a nw DriveTrainMotionControlSimplifiled, and had to change
several objects (Encoders, Drivetrain, gyro) as public static. Although
this goes agains how I program at work, being the amount of code for the
roboti is much smaller and the robot will only have one gyro, one
robotdrive this is a proper (or at least reasonable) way to code on a
robot.