Code to do a Motion controlled Arch turn.
This controls the angular speed to be proportional to the speed of the
robot.  I think instead want it to track the correct angle based on
where we are at in the path, that way it would catchup incase it was
disturbed along the way.  Currently if gets knocked along the way it
will not adjust, but instead will get be at the wrong angle at the end.