switch to vision target width for distance
Found pixle for height was only 1 or 2 pixles so no resolution, switched
to vision target width and got 5 pixles, but the width would vary from 1
to 5, very irratic. Need to look at with at what vision is doing since
the vision align is also inconsistent based on what I saw and heard from
Colin and Michael on Saturday.