Getting Encoder controlled Shooter in place
Changed RobotSensorMap so it properly converts Encoder to RPM of the
shooter.  LidarLiteSensor has isWorking method that will be used along
with not getting 0 measure to set the RPM, if that does not work vision
based width will get a distance measurment.  If not Visio target is in
sight will pull RPM from dashboard, which will default to 1500 if no
value is available.  Shooter Adjusted pid settings to make it twice as
aggresive.