Created a RobotConstants class and 3 instances LilGeek, Mackinac & Rudy
The class tries to get from the dashboard what Robot it is on and
defaults to Mackinac. Also modified SensorMap to use the class for the
encoder distance per click and DriveTraine for its PID values.  Need to
prove it works and then use it for other PID settings and anything else
it would be useful for.