Removed limitation on Power in Motion Control
Motion Control should have full power available to it to accomplish what
it is being request to do.  Also in MotionController put the PID setting
back to what rudy had.  There was a value of 0.9 which ment that if
rotation was set to 1RPM that would be 6 deg/sec so 90% power would
apply when just running at 5 Deg/sec, this is way to sensetive.  It ment
the power limiter was what was controlling the rotation, not the PID
motion control.