Code ready to test for TalonSRX on Shooter for RPM control
The Encoder was removed from Robot Sensor Map, since it no longer would
talk directly to the RoboRio.  Instead it is defined in the Shooter
Object which deals with the TalonSRX, which is told he Magnetic encoder
is hooked to it.  
RobotControllerMap updated, no longer uses PWM instead using CANTalon
with ID 1 (which the TalonSRX was configured to be through the RoboRio
website).
Also Robot was updated to no longer pass the encoder to the shooter,
since all the shooter object has now is the Talon SRX.  code compiles,
still needs to be tested. Need to update TalonSRX to latest firmware, do
though RoboRio Website, need to login to the admin account.