Fixed shooter so RPM sign matches request
Found there was another place we were inverting the motor, and we had
been using a differant command to invert the Talon SRX behaviour.
getting m_Shooter.reverseOutput(true); removed which seem to not reverse
the motor but just changed the reported RPM out and getting the
m_Shooter.setInverted(false); fixed things.